feat: ListableInternshipOffer refacto , tests to be completed
