+++ app/views/dashboard/internship_offers/_tabs.html.slim
@@ -1,4 +1,4 @@
-- forwarded_params = params.permit(:latitude, :longitude, :radius, :city, :order, :direction)
+- forwarded_params = params.permit(:latitude, :longitude, :radius, :city, :order, :direction, :keyword)
 
 ul.nav.nav-tabs.test-dashboard-nav.bb-1.bc-light.col-12.nav-xs-inline
   - is_active = params[:filter].blank?
